Refactors github team table to its own component
